moved the character to its own file
